Fix server component error
