+++ apps/web/app/(app)/[emailAccountId]/assistant/Rules.tsx
@@ -66,7 +66,7 @@ import { DEFAULT_COLD_EMAIL_PROMPT } from "@/utils/cold-email/prompt";
 import {
   STEP_KEYS,
   getStepNumber,
-} from "@/app/(app)/[emailAccountId]/onboarding/OnboardingContent";
+} from "@/app/(app)/[emailAccountId]/onboarding/steps";
 
 export function Rules({
   showAddRuleButton = true,
