+++ apps/web/app/(app)/[emailAccountId]/onboarding/OnboardingContent.tsx
@@ -22,33 +22,10 @@ import { useSignUpEvent } from "@/hooks/useSignupEvent";
 import { isDefined } from "@/utils/types";
 import { StepCompanySize } from "@/app/(app)/[emailAccountId]/onboarding/StepCompanySize";
 import { usePremium } from "@/components/PremiumAlert";
-
-export const STEP_KEYS = {
-  INTRO: "intro",
-  FEATURES: "features",
-  WHO: "who",
-  COMPANY_SIZE: "companySize",
-  LABELS: "labels",
-  DRAFT: "draft",
-  CUSTOM_RULES: "customRules",
-} as const;
-
-const STEP_ORDER = [
-  STEP_KEYS.INTRO,
-  STEP_KEYS.FEATURES,
-  STEP_KEYS.WHO,
-  STEP_KEYS.COMPANY_SIZE,
-  STEP_KEYS.LABELS,
-  STEP_KEYS.DRAFT,
-  STEP_KEYS.CUSTOM_RULES,
-] as const;
-
-export function getStepNumber(
-  stepKey: (typeof STEP_KEYS)[keyof typeof STEP_KEYS],
-): number {
-  const index = STEP_ORDER.indexOf(stepKey);
-  return index === -1 ? 1 : index + 1;
-}
+import {
+  STEP_KEYS,
+  STEP_ORDER,
+} from "@/app/(app)/[emailAccountId]/onboarding/steps";
 
 interface OnboardingContentProps {
   step: number;
