+++ version.txt
@@ -1,1 +1,1 @@
-v2.20.26
+v2.20.27
